# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jeremiah Generation Youth API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: public_only
          in: query
          schema:
            type: boolean
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (leader/admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (leader/admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /events/{id}/stats:
    get:
      operationId: getEventStats
      tags: [events]
      summary: Get attendance and RSVP stats for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStats"

  # Attendance
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: session_date
          in: query
          schema:
            type: string
            format: date
        - name: event_id
          in: query
          schema:
            type: string
        - name: profile_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in a person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "201":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
...
  /attendance/checkin-by-name:
    post:
      operationId: checkInByName
      tags: [attendance]
      summary: Self check-in by name or phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelfCheckInInput"
      responses:
        "201":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # RSVPs
...
  /checkin/search:
    get:
      operationId: searchForCheckIn
      tags: [checkin]
      summary: Search members/visitors by name or phone for check-in
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"

  # QR Codes
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent check-ins and registrations
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/attendance-history:
    get:
      operationId: getAttendanceHistory
      tags: [dashboard]
      summary: Get attendance per session/date (all-time records)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Attendance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceHistoryEntry"
components:
  schemas:
...
    Profile:
      type: object
      required: [id, full_name, role, created_at]
      properties:
        id:
          type: string
        clerk_id:
          type: ["string", "null"]
        full_name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [male, female, other, "null"]
        age:
          type: ["integer", "null"]
        heard_from:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, leader, member, visitor]
        created_at:
          type: string
          format: date-time
...
    Event:
      type: object
      required: [id, title, date, time, location, is_public, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
          format: date
        time:
          type: string
        location:
          type: string
        created_by:
          type: ["string", "null"]
        age_min:
          type: ["integer", "null"]
        age_max:
          type: ["integer", "null"]
        custom_requirements:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/EventRequirement"
        is_public:
          type: boolean
        rsvp_count:
          type: ["integer", "null"]
        attendance_count:
          type: ["integer", "null"]
        created_at:
          type: string
          format: date-time
    EventRequirement:
      type: object
      required: [label, required]
      properties:
        label:
          type: string
        required:
          type: boolean
    EventInput:
      type: object
      required: [title, date, time, location]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        location:
          type: string
        age_min:
          type: ["integer", "null"]
        age_max:
          type: ["integer", "null"]
        custom_requirements:
          type: array
          items:
            $ref: "#/components/schemas/EventRequirement"
        is_public:
          type: boolean
          default: true
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        location:
          type: string
        age_min:
          type: ["integer", "null"]
        age_max:
          type: ["integer", "null"]
        custom_requirements:
          type: array
          items:
            $ref: "#/components/schemas/EventRequirement"
        is_public:
          type: boolean
    EventStats:
      type: object
      required: [event_id, rsvp_count, going_count, not_going_count, maybe_count, attendance_count]
      properties:
        event_id:
          type: string
        rsvp_count:
          type: integer
        going_count:
          type: integer
        not_going_count:
          type: integer
        maybe_count:
          type: integer
        attendance_count:
          type: integer
    AttendanceRecord:
      type: object
      required: [id, session_date, checked_in_at, check_in_method]
      properties:
        id:
          type: string
        profile_id:
          type: ["string", "null"]
        event_id:
          type: ["string", "null"]
        checked_in_at:
          type: string
          format: date-time
        session_date:
          type: string
          format: date
        check_in_method:
          type: string
          enum: [manual, self, qr]
        profile:
          $ref: "#/components/schemas/Profile"
    CheckInInput:
      type: object
      required: [profile_id, check_in_method]
      properties:
        profile_id:
          type: string
        event_id:
          type: ["string", "null"]
        check_in_method:
          type: string
          enum: [manual, self, qr]
    SelfCheckInInput:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
...
    ActivityItem:
      type: object
      required: [type, description, timestamp]
      properties:
        type:
          type: string
          enum: [check_in, registration, membership_request, rsvp]
        description:
          type: string
        profile_name:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
    AttendanceHistoryEntry:
      type: object
      required: [session_date, total_count, member_count, visitor_count]
      properties:
        session_date:
          type: string
          format: date
        total_count:
          type: integer
        member_count:
          type: integer
        visitor_count:
          type: integer
        event_title:
          type: ["string", "null"]