# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeremiah Generation Youth API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  type: ActivityItemType;
  description: string;
  /** @nullable */
  profile_name?: string | null;
  timestamp: Date;
}