# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeremiah Generation Youth API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileGender } from "./profileGender";
import type { ProfileRole } from "./profileRole";

export interface Profile {
  id: string;
  /** @nullable */
  clerk_id?: string | null;
  full_name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  gender?: ProfileGender;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  heard_from?: string | null;
  role: ProfileRole;
  created_at: Date;
}